a function that executes a function given as a parameter on each element of an array

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,4 +1,6 @@
 #include "function_pointers.h"
+#include <stdio.h>
+#include <stddef.h>
 
 /**
  * array_iterator - executes a function give
@@ -12,12 +14,12 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t i;
+	unsigned int  i;
 
 	if (array == NULL || action == NULL)
 		return;
 
-		for (i = 0; i < size; i++)
+	for (i = 0; i < size; i++)
 		{
 			action(array[i]);
 		}
